+++ internet_speed_test_extension/Program.cs
@@ -9,7 +9,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace internet_speed_test_extension;
+namespace SpeedTest.Extension;
 
 public class Program
 {
@@ -25,8 +25,8 @@ public class Program
             // We are instantiating an extension instance once above, and returning it every time the callback in RegisterExtension below is called.
             // This makes sure that only one instance of SampleExtension is alive, which is returned every time the host asks for the IExtension object.
             // If you want to instantiate a new instance each time the host asks, create the new instance inside the delegate.
-            internet_speed_test_extension extensionInstance = new(extensionDisposedEvent);
-            server.RegisterClass<internet_speed_test_extension, IExtension>(() => extensionInstance);
+            SpeedTestExtension extensionInstance = new(extensionDisposedEvent);
+            server.RegisterClass<SpeedTestExtension, IExtension>(() => extensionInstance);
             server.Start();
             
             // This will make the main thread wait until the event is signalled by the extension class.